refactor: 디렉토리 재설정
