+++ tomcat/src/main/java/org/apache/coyote/util/StaticResourcePathGenerator.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.response;
+package org.apache.coyote.util;
 
 public class StaticResourcePathGenerator {
